requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
disnake==2.9.2
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.3
idna==3.8
multidict==6.0.4
pipdeptree==2.23.1
pycodestyle==2.12.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
tomli==2.0.1
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.9.4
